Misc bugfixes reported by Clinton
logging could fail if given an excessively long message. This happened
primarily with the event key lists, which now use log_list() which takes
the message template and keys and splits into multiple sub-1024
character messages as needed.

do_shell() attempted to pass all of its parameters to subprocess.call()
and broke if there was anything which didn't automatically stringify. It
now looks only for two keys ("key" and "info") and sets them as
CRANKD_{key.upper()} to avoid conflicts - we'll probably want to revise
this list to see which other values are useful for simple scripts.
Python handlers, of course, still have access to everything in native
form.

Added a -d/--debug option to set the default logging level and made the
default file-based log level use the same value as stdout